add writeTimes()

+++ include/Writer.hpp
@@ -36,5 +36,5 @@ public:
 
     void writeClusters(std::string filepath) override;
     void writeClustering(std::string filepath) override;
-    void writeTimes(std::vector<float> time, std::string);
+    void writeTimes(std::vector<double> time, std::string);
 };
